# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: InterviewOS API
  version: 1.0.0
  summary: "REST API for InterviewOS — the interview-prep learning operating system (GA track: Backend SDE3)."
  description: |
    REST API under `/api/v1`. JWT bearer access tokens (short-lived) + rotating
    refresh tokens. All field names and enum values are consistent with
    `04-DATABASE-SCHEMA.md` and `05-API-CONTRACTS.md`.
  contact:
    name: InterviewOS Engineering
    email: eng@interviewos.app
  license:
    name: Proprietary
servers:
  - url: https://api.interviewos.app/api/v1
    description: Production
  - url: https://staging-api.interviewos.app/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /problems:
    get:
      tags: [DSA]
      summary: List DSA problems (filter by company, difficulty, pattern)
      operationId: listProblems
      parameters:
        - { name: difficulty, in: query, schema: { $ref: '#/components/schemas/Difficulty' } }
        - { name: pattern_id, in: query, schema: { type: string, format: uuid } }
        - { name: topic_id, in: query, schema: { type: string, format: uuid } }
        - { name: company_id, in: query, schema: { type: string, format: uuid } }
        - { name: source, in: query, schema: { $ref: '#/components/schemas/ProblemSourceName' } }
        - { name: solved, in: query, schema: { type: boolean } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Problem' } }
  /problems/{problemId}:
    get:
      tags: [DSA]
      summary: Get a problem with patterns, sources, company frequency, progress
      operationId: getProblem
      parameters:
        - { name: problemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /design-problems:
    get:
      tags: [SystemDesign]
      summary: List HLD design problems (ordered catalog)
      operationId: listDesignProblems
      parameters:
        - { name: difficulty, in: query, schema: { $ref: '#/components/schemas/Difficulty' } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/DesignProblem' } }
  /design-problems/{designProblemId}:
    get:
      tags: [SystemDesign]
      summary: Get a design problem with all sections
      operationId: getDesignProblem
      parameters:
        - { name: designProblemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignProblemDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /design-problems/{designProblemId}/progress:
    get:
      tags: [SystemDesign]
      summary: Get the caller's progress on a design problem
      operationId: getDesignProblemProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: designProblemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK (status is "not_started" when none recorded)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignProblemProgress' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [SystemDesign]
      summary: Record progress on a design problem (status + confidence)
      operationId: saveDesignProblemProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: designProblemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [not_started, in_progress, completed, needs_review] }
                confidence: { type: integer, minimum: 1, maximum: 5, nullable: true }
                time_spent_minutes: { type: integer, minimum: 0 }
                notes: { type: string, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignProblemProgress' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [SystemDesign]
      summary: Clear the caller's progress on a design problem
      operationId: deleteDesignProblemProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: designProblemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: No Content (progress cleared; no-op when none recorded) }
        '404': { $ref: '#/components/responses/NotFound' }
  /problems/{problemId}/progress:
    get:
      tags: [DSA]
      summary: Get the caller's progress + stored solution for a DSA problem
      operationId: getProblemProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: problemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ProblemProgress' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [DSA]
      summary: Record solve state + solution for a DSA problem
      operationId: saveProblemProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: problemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solved: { type: boolean }
                confidence: { type: integer, minimum: 1, maximum: 5, nullable: true }
                time_spent_minutes: { type: integer, minimum: 0 }
                solution_code: { type: string, nullable: true }
                solution_language: { type: string, nullable: true }
                solution_notes: { type: string, nullable: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ProblemProgress' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [DSA]
      summary: Clear the caller's progress + solution for a DSA problem
      operationId: deleteProblemProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: problemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: No Content }
  /problems/solved:
    get:
      tags: [DSA]
      summary: The caller's solved/attempted problem log
      operationId: listSolvedProblems
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ProblemProgress' } }

  # ---------------- BACKEND ENGINEERING ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token in `Authorization: Bearer <token>`."
    refreshAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating refresh token (HttpOnly cookie; may also be sent in body).
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      description: "Comma-separated fields, `-` prefix for desc. e.g. `-created_at,title`."
      schema: { type: string }
...
    Query:
      name: q
      in: query
      description: Full-text search query.
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    ValidationError:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
...
    ProblemProgress:
      type: object
      properties:
        problem_id: { type: string, format: uuid }
        status: { type: string, enum: [not_started, in_progress, completed, needs_review] }
        solved: { type: boolean }
        confidence: { type: integer, nullable: true }
        attempts: { type: integer }
        time_spent_minutes: { type: integer }
        solved_at: { type: string, format: date-time, nullable: true }
        solution_code: { type: string, nullable: true }
        solution_language: { type: string, nullable: true }
        solution_notes: { type: string, nullable: true }
        solution_updated_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time }
    DesignProblemProgress:
      type: object
      properties:
        design_problem_id: { type: string, format: uuid }
        status: { type: string, enum: [not_started, in_progress, completed, needs_review] }
        confidence: { type: integer, nullable: true }
        attempts: { type: integer }
        time_spent_minutes: { type: integer }
        notes: { type: string, nullable: true }
        first_completed_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time }
    # ---- Enums ----
...
    Difficulty: { type: string, enum: [easy, medium, hard] }
...
    ProgressStatus: { type: string, enum: [not_started, in_progress, completed, needs_review] }
    ConfidenceLevel: { type: integer, minimum: 1, maximum: 5 }
...
    ProblemSourceName: { type: string, enum: [blind75, neetcode150, grind75, tech_interview_handbook, leetcode_top, striver_sde, custom] }
    ProblemPlatform: { type: string, enum: [leetcode, hackerrank, codeforces, interviewbit, gfg, custom] }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Stable machine code, e.g. VALIDATION_ERROR }
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    PaginationMeta:
      type: object
      required: [page, page_size, total, total_pages]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    PaginatedEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # ---- Auth ----
...
    Problem:
      type: object
      required: [id, track_id, slug, title, difficulty]
      properties:
        id: { type: string, format: uuid }
        track_id: { type: string, format: uuid }
        topic_id: { type: [string, 'null'], format: uuid }
        slug: { type: string }
        title: { type: string }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        platform: { $ref: '#/components/schemas/ProblemPlatform' }
        external_id: { type: [string, 'null'] }
        url: { type: [string, 'null'], format: uri }
        estimated_minutes: { type: integer }
        frequency_score: { type: number }
        is_premium: { type: boolean }
    ProblemDetail:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            prompt_summary: { type: [string, 'null'] }
            approach_md: { type: [string, 'null'] }
            common_mistakes: { type: [string, 'null'] }
            patterns: { type: array, items: { $ref: '#/components/schemas/Pattern' } }
            sources: { type: array, items: { $ref: '#/components/schemas/ProblemSource' } }
            company_frequency: { type: array, items: { $ref: '#/components/schemas/ProblemCompanyFrequency' } }
            progress: { oneOf: [ { $ref: '#/components/schemas/UserProblemProgress' }, { type: 'null' } ] }
    Pattern:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: string, format: uuid }
        track_id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        when_to_use: { type: [string, 'null'] }
    ProblemSource:
      type: object
      required: [source]
      properties:
        source: { $ref: '#/components/schemas/ProblemSourceName' }
        source_rank: { type: [integer, 'null'] }
        source_url: { type: [string, 'null'], format: uri }
    ProblemCompanyFrequency:
      type: object
      required: [company_id, frequency]
      properties:
        company_id: { type: string, format: uuid }
        company_name: { type: string }
        frequency: { type: number }
        last_seen_period: { type: [string, 'null'] }

    # ---- Design / LLD ----
    DesignProblem:
      type: object
      required: [id, slug, title, difficulty, order_index]
      properties:
        id: { type: string, format: uuid }
        track_id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        order_index: { type: integer }
    DesignProblemDetail:
      allOf:
        - $ref: '#/components/schemas/DesignProblem'
        - type: object
          properties:
            requirements_md: { type: [string, 'null'] }
            capacity_estimation_md: { type: [string, 'null'] }
            api_design_md: { type: [string, 'null'] }
            data_model_md: { type: [string, 'null'] }
            high_level_design_md: { type: [string, 'null'] }
            caching_md: { type: [string, 'null'] }
            queueing_md: { type: [string, 'null'] }
            scaling_md: { type: [string, 'null'] }
            tradeoffs_md: { type: [string, 'null'] }
            failure_handling_md: { type: [string, 'null'] }
            alternatives_md: { type: [string, 'null'] }
            interview_tips_md: { type: [string, 'null'] }
            follow_up_questions: { type: array, items: { type: string } }
...
    UserProblemProgress:
      type: object
      required: [id, user_id, problem_id, status, solved]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        problem_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/ProgressStatus' }
        confidence: { oneOf: [ { $ref: '#/components/schemas/ConfidenceLevel' }, { type: 'null' } ] }
        attempts: { type: integer }
        solved: { type: boolean }
        time_spent_minutes: { type: integer }
        last_attempt_at: { type: [string, 'null'], format: date-time }
        solved_at: { type: [string, 'null'], format: date-time }
        notes: { type: [string, 'null'] }

    # ---- Revision ----
...